+++ functional_tests/example_cre.R
@@ -1,12 +1,12 @@
 set.seed(2022)
 
 # Set Experiment Parameter
-n_rules <- 4
+n_rules <- 2
 sample_size <- 2000
 effect_size <- 2
 confounding <- "no"
-ite_estimator <- "bart"
-pfer <- 0.5#n_rules/(effect_size+1)
+ite_estimator <- "aipw"
+pfer <- n_rules/(effect_size+1)
 
 # Set Method and Hyper Parameters
 method_params <- list(ratio_dis = 0.5,
@@ -19,15 +19,16 @@ hyper_params <- list(intervention_vars = NULL,
                      ntrees_rf = 40,
                      ntrees_gbm = 40,
                      node_size = 20,
-                     max_nodes = 8,
-                     max_depth = 3,
+                     max_nodes = 4,
+                     max_depth = 2,
                      t_decay = 0.025,
                      t_ext = 0.01,
                      t_corr = 1,
-                     t_pvalue = 0.05,
                      stability_selection = TRUE,
                      cutoff = 0.8,
-                     pfer = pfer)
+                     pfer = pfer,
+                     B = 20,
+                     subsample = 0.2)
 
 # Set Ground Truth
 {