+++ tests/testthat/test-discover_rules.R
@@ -23,10 +23,11 @@ test_that("discover_rules works as expected!", {
                        t_decay = 0.025,
                        t_ext = 0.025,
                        t_corr = 1,
-                       t_pvalue = 0.05,
                        stability_selection = TRUE,
                        cutoff = 0.8,
-                       pfer = 1)
+                       pfer = 1,
+                       B = 2,
+                       subsample = 0.5)
 
   # Input checks
   check_input_data(y = y, z = z, X = X)